+++ SleekFlow.Todo.Domain/TodoCreatedEvent.cs
@@ -3,7 +3,7 @@
 public class TodoCreatedEvent : IEvent
 {
     public Guid Id { get; set; }
-    public int EventNumber { get; set; }
+    public long EventNumber { get; set; }
     public string RaisedBy { get; set; } = null!;
     public DateTime RaisedAt { get; set; }
 }
